backend/pyproject.toml
[tool.poetry]
name = "brain diagnosis"
version = "0.1.0"
description = ""
authors = ["Artachtron <florent.vaucher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<=3.10"
tensorflow-io-gcs-filesystem = "< 0.32.0"
fastapi = "^0.111.0"
pillow = "^10.3.0"
numpy = "^1.26.4"
keras = "2.13.1"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
